+++ src/main/java/com/knu/ntttt_server/token/repository/UserGachaTokenRepository.java
@@ -11,5 +11,5 @@ import java.util.Optional;
 public interface UserGachaTokenRepository extends JpaRepository<UserGachaToken, Long> {
     Optional<UserGachaToken> findByUserAndDate(User user, LocalDate date);
 
-    Integer countByToken_Id(Long tokenId);
+    Long countByToken_Id(Long tokenId);
 }
