+++ src/main/java/com/knu/ntttt_server/token/service/GachaService.java
@@ -5,5 +5,5 @@ import com.knu.ntttt_server.token.dto.GachaDto;
 public interface GachaService {
     GachaDto.GachaRes getGachaToken(String username);
     GachaDto.GachaRes playGacha(String username);
-    Integer getUserNumberSeeingSameToken(Long tokenId);
+    Long getUserNumberSeeingSameToken(Long tokenId);
 }
